Change the output directory of the temp log file
This commit change the command "journal > logs" to output the "logs"
file under /var/logs/journal instead.

journal > logs will try to output the file under /usr/bin, which is by
convention not a correct  place to write dynamic files  into. In some
system, the directory is expected to be locked as Read-Only. This can
cause the download log  feature on IED malfunctioning.

--transform -C are used for compatibility that keeps the archive content
the same.